File Deletion using nonce, file renaming, examples for custom-name functions,
If config.uploadDeleteLink is false then file deletion is only possible for logged in user. If it's true then any user is allowed to remove if they provide correct nonce.
Also only if that setting is true the deleteUrl value is added to returned url on upload that informs uploader what deletionUrl is.
Also added success page to offset error page. Provided some basic custom name functions for end user to use.